refactor(V1__Init): refactoring migration

+++ flyway/sql/V1__Init.sql
@@ -5,7 +5,7 @@ CREATE TABLE bank.BALANCE (
   balance REAL NOT NULL
 );
 
-CREATE FUNCTION bank.TRANSFER(from_user TEXT, to_user TEXT, amount REAL) RETURNS void AS $$
+CREATE FUNCTION bank.TRANSFER(from_user TEXT, to_user TEXT, amount BIGINT) RETURNS void AS $$
 BEGIN
   UPDATE bank.BALANCE SET balance = balance + amount WHERE user_id = to_user;
   UPDATE bank.BALANCE SET balance = balance - amount WHERE user_id = from_user;
